Remove incorrect dispatch_once usage
Both of the uses of dispatch_once were complete no-ops, as the `var onceToken` is being created each time the function was entered.

In the case of the regular expression, we instead create a private global with the regex, which is thread-safe. We would do it in the extension, but swift does not support static properies in extensions.

in the case of the `emojiAliases` function, we can just remove it since the function was giving us no extra behaviour.